Use DBLoadBalancerFactory to rollback all DBs, add more logging
Per conversation with Aaron, getDBLoadBalancerFactory()->rollbackMasterChanges()
should be sufficient to rollback both the Flow and master databases.

The $wikiDbw and $flowDbw params have been removed from the DeferredUpdate to
further simplify the code execution. Since we are executing the code in a try/
catch block and then handling the rollback, we shouldn't need DeferredUpdates
to also potentially intervene with halting the update when a DB is rolled back.

Add logging in the section of SubmissionHandler that checks permissions, to
verify which user is being used in the permissions check and also to get a
complete picture of all errors, as later on when we iterate over
$block->getErrors(), we only throw an exception for the first item in the array,
and consequently some errors might remain hidden from the logs.

Bug: T70526
Change-Id: I38e07dcee06400e769c6191368a21e7d87b09fec