WIP: Add job for fixing boards when InvalidDataException is thrown
This doesn't fix the underlying problem of why blocked users in some
circumstances cannot disable Flow on their talk page, but at least it fixes
those pages which are in a broken state.

- Add class for fixing inconsistent boards, using code from maintenance script
used for that purpose
- Add job for fixing boards when InvalidDataException is thrown
- Refactor maintenance script to use shared class

TODO/Questions:

- Could we end up in a state where jobs are endlessly enqueued with no hope of
completion? (If so, do we need to do something else to clean up those cases?)
- Instead of trying to fix these pages on failure, would we be better off with
checking permissions before the user attempts to move in the first place?
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/Flow/+/455897
- Since we're enqueueing a job that could fix the problem, should we redirect
the user somewhere and show them a message rather than allowing the exception
to bubble up?
- Untested code

Bug: T70526
Change-Id: I19c0a0cfea3051eac66a658f5d10a6c14ee87d96