Fix recentchanges event not emitted for topic on empty board
If in the context of a POST request, check against the master DB. This is
important for recentchanges actions; if a user posts a topic on an empty flow
board then querying the replica results in $this->exists getting set to
false. Querying the master DB correctly returns that the title exists, and the
recent changes event can propagate.

Bug: T187861
Change-Id: If763693982c142dfa63c4edcc4e090bece50addc