Make ContributionsQuery work with a value object instead of a ContribsPager
This is an alternative proposal for the issue described in T212771. With
this and teh depending patch I827e0fa, the code does not need access to the
(now private) properties in the ContribsPager class any more. It doesn't even
need to know if the object it gets is a ContribsPager or a generic value
object, holding nothing but the options the code actually needs.

See I827e0fa for the related core change.

Bug: T212771
Change-Id: I211d2cdc0268c7995a098233dfa614047719082d