Update patch set 2
Patch Set 2: Code-Review-1

Wait, so it's not OK for Flow to directly access properties on a ContribsPager, but it is OK to directly access those same properties on a DeletedContribsPager? Why is that? Shouldn't we either use a value object for both of them, or a Pager object for both of them?

Patch-set: 2
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1