Make ContributionsQuery accept value objects instead of DeletedContribsPager
This goes along with I211d2cd and is effectively the second part of this patch.
It makes the code consistent in so far that both hook handlers now behave the
same again. I intentionally uploaded this as a separate patch to make the two
smaller and easier to review.

Note the two patches here in the Flow extension are cross-compatible with and
without the core patch I827e0fa being applied or not. What the two patches here
do is making this code more flexible. It does not need the precise
(Deleted)ContribsPager objects any more, but accepts any object with the
expected properties instead.

Bug: T212771
Change-Id: I3687b23e6d92d6cf5cde8686b9fbf0a3cb56a9a3