Remove all default return true from hook handler functions
This is the default anyway for a long time now. There is no difference
between returning true and returning nothing.

Note that many of the other hook handlers this patch currently does not
touch can return false. I believe this is wrong in almost all situations.
Returning false means this will stop all *other* hook handlers (on the
same hook) from being executed. This is typically not what needs to be
done. When there is an error in *this* hook handler, it should stop and
return nothing (or true). But it does not need to stop *other* handlers
in other extensions.

But this is for a later patch.

Change-Id: I86df4e0ac0c464f3d634232efa5281a26028a573