Move indexField names into (Deleted)ContributionsRow sub-classes
This avoids hard-coding the string names in a place where they don't
really belong. Instead, the field names are now exactly where they are
going to be used anyway: in the two FormatterRow sub-classes.

Feel free to update this patch however you like. For example, instead
of a getter it could use named constants.

Change-Id: I949a576af66736494d2ecf5024f53d728cdf7a8c