WIP: Store full HTML documents for Flow content
Add an API param for "body_only" which defaults to "false". If set to true, the
Utils::convert() method will function as before and return the body of an HTML
document. This is used when rendering a flow reply in VE mode. If set to false,
the full HTML document will be stored in the database.

Bug: T209114
Change-Id: Ic1168ff843cbf792e290c4828b6191ab466a3c47