Update patch set 2
Patch Set 2:

I think this will cause the client to still receive body-only HTML from old storage, without knowing to body-wrap it. If that happens, it'd probably be best for the server to correct that and always serve a full document. I'll poke around more and verify that this is the case.

Patch-set: 2
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>