Update patch set 2
Patch Set 2:

So, this generally doesn't break, and getting it to break is pretty convoluted and edge-case-reliant, but it still doesn't make me feel great.

Here's how things can go wrong:
1. Save a comment while running the old code
2. Switch to the new code
3. Edit the comment in VE (meaning, make sure VE is your default editor, then start editing the comment, so it loads straight into VE)
4. The HTML of the comment is requested from api.php; response is the HTML as stored in the DB, without a <body> wrapper
5. This HTML is loaded straight into VE, without first wrapping it in a <body> tag, because this commit removes the wrapping code
6. This can cause <link> and <meta> tags that are supposed to be in the body to be reparented into the <head>, and subsequently lost (you'd expect it to be saved in the <head> of the now-full-document comment in the DB, but it isn't, likely because we only send the <body> to the server when saving)

Admittedly, this is an edge case of an edge case, because the kinds of things that Parsoid represents with <link>/<meta> tags are rare in Flow comments (they're things like categories and double-underscore magic words). Nevertheless, I don't feel great about api.php sometimes serving a full document and sometimes serving just a body when asked for HTML, so it would be nice if that could be addressed.

Patch-set: 2
Reviewer: Catrope <5@e9e9afe9-4712-486d-8885-f54b72dd1951>