mw.flow.ve.Target: Change surface #setDisabled to #setReadOnly
Surfaces can no longer be disabled, but instead they can be set
to read-only. The difference is that content in read-only surfaces
can be selected and inspected.

It seems that Flow only uses this when trying to add a new topic
(mw.flow.ui.NewTopicWidget) on an uneditable page (e.g. protected).
Since there is no content in the surface in this case, the effect
of old disabled mode and new read-only mode is nearly the same
(except for a blinking cursor and focus styling).

Depends-On: Ifaa74be266c048d87f94b4ae2df59cafc7aa155f
Change-Id: I4771dcd59368284c105989224507bf8f536efaea