Don't override message and stack trace for UnknownWorkflowIdException
There are two different types of this exception, and swallowing the
message makes it hard to figure out which one is being thrown. For
exceptions like InvalidTopicUuidException it makes sense to have a nice
error message because users can trigger it directly by going to an
invalid URL. But UnknownWorkflowIdException can only be thrown if
there's bad data in the database, so it's OK to show an uglier page with
a stack trace in that case.

Bug: T217330
Change-Id: I65bda6b215d2b59b7386c4c2961d522e8507c9bf