Conversion\Utils: Work around DOMDocument serialization bug
When the following HTML:
    <p data-foo="&quot;&lt;!-- foo&amp;bar --&gt;">Foo&amp;Bar</p>
is parsed with DOMDocument, then serialized with ->saveHTML(), you get:
    <p data-foo='"<!-- foo&bar -->'>Foo&amp;Bar</p>
which is invalid (unencoded ampersand) and causes errors when parsed
a second time. This only happens when the attribute value contains
something that looks like an HTML comment.

This doesn't just cause errors but can also cause data corruption. If
the attribute value starts as "&lt;!-- foo &amp;eacute; bar--&gt", it'll
change to "<!-- foo &eacute; bar-->" when round-tripped once, then to
"<!-- foo é bar -->" when round-tripped a second time.

To work around this bug, serialize our HTML DOMs using saveXML()
instead. We need to pass the LIBXML_NOEMPTYTAG flag to prevent empty
tags from being serialized as self-closing tags (e.g. <a href="xyz"></a>
would become <a href="xyz" />, which is not valid HTML). With that flag,
the output is semantically identical to saveHTML(); there is a syntactic
difference where saveHTML() uses single quotes for attribute values that
contain double quotes, whereas saveXML() uses double quotes and &quot;
encoding for those.

Bug: T217766
Change-Id: I48e7cecddc42c2185d5a51074ef21aca58010b66