Update patch set 6
Patch Set 6: Code-Review+2

Going ahead for now, as it was still seen and causing core commits to fail. I'll leave it to the team to investigate and decide how to move forward if/when this is a priority.

Patch-set: 6
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2