Add maintenance script to reserialize revision content
This script finds revisions whose content is not at the latest Parsoid
version, and converts them to wikitext then back to HTML so that they
are at the latest Parsoid version.

In order to update the contents of existing revisions in place, the
script needs to do some reflection hackery to make rev_content and
friends writable, then call the non-public setContentRaw() method
(broken out of setContent()).

Bonus: convert UUIDs to alphadecimal format before comparing them in
FeatureIndex, so that the output of the maintenance script isn't drowned
out by warnings about UUID::__toString() being used.

Bug: T209120
Change-Id: I7491296558d596453b8680185667483beaad4c6d