Use $IP in maintenance where available
This means when our maintenance scripts are called by update.php or
anything else that's already initialized a Maintenance object, we'll get
the correct path without having to hope that the extension is checked
out in a subdirectory of the install path.

The .phpcs.xml change should not be necessary IMO. See comment on
T104501.

Change-Id: I3e2699c5464f1922e66100d3217d4883d2543474