Replace calls to deprecated MWTimestamp::getHumanTimestamp()
Deprecated since MediaWiki 1.26.

The method UUID::getHumanTimestamp() is a little weird. It was
introduced via Id5af3e9 in 2013, and used in two template files. The
templating got rewritten in 2014. The method is unused since then. I
could not find any usage. Definitely not in this codebase. But also not
in other codebases (which should not be a thing anyway):
https://codesearch.wmflabs.org/search/?q=%3EgetHumanTimestamp%5C(&i=1
All methods that show up in this search either refer to the MWTimestamp
or Language class.

Change-Id: I8b03c4984ebcefdd36ab4fec4c979ca666a8b76b