Speculative fix for undefined index error in ChangesListFormatter
I haven't been able to reproduce this bug, and it appears to be
extremely rare in production (once in the past 30 days), but from static
analysis it looks like getTimestampLink() should call
setIncludeHistoryProperties( true ), otherwise $data['dateFormats'] is
an empty array. However, in practice, it seems like format() is almost
always called before getTimestampLink() is called, and format() also
enables history properties. So I haven't managed to get this to break in
practice, but the getTimestampLink() function should at least be
theoretically correct.

Bug: T205191
Change-Id: I440e360fe30c80e494360c51c8b2217bba1f6377