Fix fatal in ChangesListFormatter::getLogTextLinks()
This function assumed that RC entries to do with a Flow page would all
be created by Flow, but that's not necessarily true. Some Flow pages
are associated with Wikidata items, in which case Wikidata RC rows get
mixed in too.

A lot of things in this function are poorly thought through and it
doesn't work very well, but this commit doesn't attempt to solve that.
It just tweaks it so that it doesn't fatal when processing a mix of Flow
and Wikidata RC entries.

Bug: T228290
Change-Id: I832296e8dc3f6a8b8ad9871fa5ea44a5641605e8