Consider invalid format a user input error
This codebase doesn't seem to have an explicit validation
layer for view actions parameters like it does for
POST actions and API calls.

Using InvalidInputException in RevisionFormatter->setContentFormat
makes it so that invalid GET URLs similar to the one specified
in the task will return a standard error page and won't spam
the log.

The drawback is that this code is not directly at the user
input level so it is conceivable that another code path
could provide an invalid format and it would be considered
a user error and ignored.

Bug: T228417
Change-Id: I767ec0a83db6dcce061aea5a0658f0698ea62e0e