Always allow archive board creation when disabling Flow beta feature
When disabling the Flow beta feature, the Flow board talk page is
moved to a subpage for future reference and can be restored when
re-enabling the beta feature.

Deep in the BoardContentHandler/TalkpageManager there is code
that checks if the current user ($wgUser) can create the archive
page. Most users don't have the 'flow-create-board' right so
it fails.

I don't know why or when this has started to break.

The proposed solution is to explicitly allow the creation of
the archive page from the OptInController.

Bug: T229795
Change-Id: I1d8b70f3e9f597cd75f2a65decd3a13c1983c4b6
(cherry picked from commit b04d9443e5ad2dcdb2fdc3e43e78e4cfb0203429)