Specify that opens a new discussion topic
I think that this clarification will give the impression that you have to click
here to initiate a conversation with the user, and will therefore define the
vague notion of "subject" that the visitor can ask himself when reading it.
I hope that adding a message that has no connection to an existing topic will
be less common after this change.

Change-Id: I1e6d69cdf5c5c0b80660661646491017c7d40896