Clean-up and simplify convertToText implementation
This was inspired by and is a direct follow-up to Ia04fcf7.

* Remove not needed StubObject::unstub().
* Replace not needed count() with straight `if ( $array )`.
* Restructure confusing implementation of ConvertToText::getAllRevisions().
I transformed the code in very small steps to be sure it still does exactly
what it was doing before.
* Turn a few large if-else into guard clauses with a single, short
if-return at the beginning of the method/loop.
* Add `array` type hints. All these places are guaranteed to not accept
anything else, because some later method call would already fail.
* Change all `protected` to `private`. There are no known subclasses, and
I don't think we should even allow such a maintenance script to have
subclasses.
* Remove spaces from array acessors like $array[$key]. As far as I
understand our coding style these spaces are only required in
constructions like [ $key => $val ], but not when accessing.

Change-Id: I223cad854a9a082f80eaad6846c592576e1b78bf