Reduce nesting by utilizing guard clauses
In the places this patch is touching I feel a guard clause does make
the code more straightforward to read.

This patch also gets rid of the array in
ArchiveNameHelper::findLatestArchiveInfo(). The only element that is
ever used is the last element (if there is one, otherwise false is
returned).

Change-Id: Ica80dfa1e21cc88ef3c1dd9ea5a793716a751620