Revert "Hooks: Use PageMoveComplete instead of TitleMoveCompleting"
This reverts commit 5f21ebd4c52f5bc9e49c96e35dedb6605dcacee2.

This broke all page moves on wikis with Flow installed. TitleMoveCompleting
happens inside the transaction, but PageMoveComplete happens outside,
and Flow was using the MoveStarting hook to start a transaction and the
MoveCompleting hook to end it. Changing these hooks broke the tranaction
assumptions, which caused misnested DB transactions, which threw a DB
error.

We'll have to either come up with a different way to do this in Flow,
or we'll need to keep a MoveCompleting hook in core, but either way,
this is an UBN so we need to revert first.

Bug: T255608
Depends-On: Id414c98161b9f560b14d4eaec8aedeec4659df27
Change-Id: I8cdef229bf0a31048ec5b69d8af0bbb0e24043d0