Pass jQuery objects into jqueryMsg
mediawiki.jqueryMsg can accept jQuery objects as message arguments and
pass them through to the parsed message, there’s no need to turn them
into HTML snippets prematurely.

Also, declare an explicit dependency on the mediawiki.jqueryMsg module,
since we rely on it. It might already be more or less guaranteed to be
loaded here, but I couldn’t fully trace the dependency graph. If it is
already pulled in transitively, then there should be no cost to the
added dependency – I believe ResourceLoader optimizes the dependency
graph in PHP before adding it to the startup module and sending it over
the wire.

Change-Id: I16757cec32f0fe638bc37c1ef55555e62f6c38c1