Fix insufficient isset() check in AbstractRevision
Before, the check was only looking for a truthy value in $params,
i.e. if it is a non-empty array. This is almost certainly the
case, as all rows in the "recentchanges" table should have at
least something in the "rc_params" field. But they don't
necessarily have a "flow-workflow-change" field.

When "flow-workflow-change" is there, it is guaranteed to contain
a "revision" field.

As described in the comment above as well as in T259739#6364407
the point of this check is to skip rows that have not been
created by Flow, i.e. rows that don't have that
"flow-workflow-change" field.

Bug: T259739
Change-Id: Ic8a90e22b8066a71826f27c3a4cd0db864c5f497