Fix metadata loading logic in CheckUserQuery
CheckUserQuery expects a third ID for Flow posts, which would be
added by the CheckUserInsertForRecentChange hook based on the
'post' rc_params field, but there is no such field (see
RecentChangesListener::onAfterInsert; also spot-checked on enwiki
that no cuc_comment field has three ids). That means the logic in
CheckUserQuery::loadMetadataBatch always ended up loading a Header;
in most cases that Header object exists but the Workflow it
references does not, so AbstractQuery::buildResult errored out
when trying to load it.

Remove the useless references to the third ID, and use the action
instead to determine the object type. Not sure if this goes to the
root of the problem (the Header object having a workflow ID that's
actually its own ID and not the ID of any Workflow seems like a bug
somewhere else in the code), but it is enough to fix the checkuser
interface.

Bug: T270334
Change-Id: I92affc7dd687b180284fe71ac81f64d371688985