container.php: merge single use listeners entries
These services are only used once, within the container definition;
this just makes them inline.

* storage.workflow.listeners -> storage.workflow
* storage.header.listeners -> storage.header
* storage.post_summary.listeners -> storage.post_summary
* storage.post.listeners -> storage.post

Update ContainerTest now that there are no separate
.listeners entries

Additionally, PostRevisionTestCase::clearExtraLifecycleHandlers()
previously altered the listeners to remove a few of them from
the container entry. Now that these are no longer in the container,
that is not possible - update to instead modify the lifecycle handlers
of the relevant ObjectManager services directly. Since we can no
longer check whether to do this based on the existance of the listener
service, hard code the list of services that should be fixed

Bug: T170330
Change-Id: Ie3440cf44d8282d018e269a251eabc6161ab0152