Change replace() calls to insert() in AbuseFilterTest::createFilter
Beyond the fact that the unique index array argument to replace() should either
be a string or array of string arrays, which is not what was provided, it is not
necessary to use replace() here when populating filters. Using INSERT will do fine,
while using REPLACE has worse performance due to different locking.

Change-Id: I5468f2feb2c422e83096d81315632b376a4f0f3e