Support TemplateStyles in Structured Discussions
Flow is doing additional processing on the Parsoid document. It was
written before RemexHtml and therefor had problems parsing.
It's using saveXML instead of saveHTML to get around some of these problems.
saveXML escapes the content of <style> tags with CDATA. Use a simple
string replacement to avoid this issue.

Bug: T252038
Change-Id: I4347e0fab88e7d95bec0f3c7fad4054b5e4d9b20