Add missing return to overridden TopicMenuSelectWidget methods
Change I1574fb8d4e (commit 013702b43e) added a `return this` to
TopicMenuSelectWidget.addItems() if the argument is not a nonempty
array, but ordinarily this method didn’t return anything, even though
its parent method in MenuSelectWidget has `return this` (for chaining).
Add the missing return statement, and for consistency do the same thing
in removeItems() as well.

Change-Id: I977b0b7ad90e588ea7b14e06789dd708b0384243