Remove ElasticSearch integration
The relevant code is not currently deployed, and given Flow's status,
it does not seem worthwhile to adapt this code to newer versions of
Elastica.

Note that the iterators are kept, as they are used in the
Bulk/Export.php code. Renaming those, moving them to a new
directory (e.g. includes/Iterators instead of includes/Search/Iterators)
and updating their documentation bcould be done in a follow-up.

Bug: T317513
Change-Id: Ief82c9cf48b714e0d0df96518ad8750436b8a967