Remove unused wgFlowUseMemcache
Flow previously used a model that tried to keep db-primary data in
perfect sync with memcached. It has sinced adopted the general
MediaWiki model of treating cache as a db-replica, populated on-demand
using WANObjectCache.

The last reference to this setting was removed in 2020 with commit
I5ea728b995e (98570f4b74) as part of T160813.

Change-Id: Ibc4316734e2350ba126dd79eb759acb2811d7110