convertToText.php: Use fresh parser for signature
Instead of trying to changing and resetting the parser, use a
clean one. This is a quick fix, probably not efficient (we could
reuse the same instance for successive getSignature() calls);
also, did not try to assess whether the reason to need an anon
option is still valid. The script was written a long time ago;
at some point we'll probably need to audit it anyway. For now,
this is enough to make it work.

Bug: T331916
Change-Id: I5bdc5a81378ae1ebfed2c13ea8dc78bacf908346