Default to using the parsoid API bundeled with MediaWiki.
This enables Mocha tests to make use of parsoid functionality.
Parsoid use is still disabled for PHPUnit tests, since outgoing
HTTP requests are blocked.

Note that this is temporary. We are working to replay VRS usage with
direct calls to parsoid in PHP over the next couple of weeks.

This patch is driven by the need to have a baseline test that works in
CI and asserts that the HTML we get from parsoid stays the same while we
are refactoring the code.

NOTE: This needs the Parsoid extension to be enabled in CI, Ic1685cf1a46

Change-Id: Ic2901f4f470e0d9e930699c72b5d0127b64985fb