Remove usage of VirtualRestService from Flow.
Remove usage to parsoid and parser functions here the two new
functions introduced as htmlToWikitext and wikitextToHTML to
perform the conversion.

Parsoid is always by default configured so isParsoidConfigured
with getServiceClient can be removed with that getVRSObject and
makeVRSObject which is creating a VRS object.

The usage of isParsoidConfigured is in onFlowAddModules where
the if condition is checking for parsoid configured, which is by
default, so the check is removed.

The isParsoidConfigured has also the reference in initFlowExtension
in Hooks.php. Here the wgFlowContentFormat is by default set to 'html'
in extension.json and parsoid is by default configured so if statement
work is removed.

In utils.php generateForwardedCookieForCli() function is only used by
makeVRSObject(), which is getting removed so it should be removed as well.

Bug: T337223
Change-Id: Ib2d7ca0e1fe1dedfb6e6bbe9e6f06856bc2d51c9