Refactor testResolveSubpageTraversal
createRelativeTitleProvider data provider is expected to provide
data not to access db and all logics should be in test fuction.

Title::newFromText() would try to access the DB and this should
not be allowed in data providers since they run first before
running the tests.

It increases the createRelativeTitle() function coverage from 75%
to 100%.

Change-Id: I7cb3109dc6035889048e2319c9f167b775e5f997