Skip affected tests unconditionally until gerrit->927619 lands
Before, Flow didn't depend on Parsoid in CI, but once we enabled
this dependency, tests that were skipped because parsoid was not
available for Flow in CI started failing.

In order not to block work moving forward (while trying to fix these
tests), let's just skip unconditionally (because that was the case
before already - so we're not making things worse).

In addition, we're already working on a patch to fix these tests when
we use Parsoid in core: Ifb85f4733be3b43b71b111df0cd3d88281101153.

Bug: T337223
Change-Id: Icd9242c3455b78fdd8eb73d42ae29ad289568837