Fix PermissionException being logged
PermissionException is an exception used for control flow, and it
should not be logged, but the isLoggable() override no longer does
anything since 4adac5bfc7a741138ad3e8055bbd3d7a3e74dfb0 (because it is
no longer a MWException).

Several places in Hooks.php already ignored it in the same way.

Bug: T356223
Change-Id: Iecb76e4d924eaf722ecfb14f2102f10f570a1cc6