Stop logging invalid topic extension and improve error page
This exception occurs when the user opens an invalid topic. Sometimes
that seems to be just a plain custom string, like "Topic:Media_options"
which is not a valid flow id, and sometimes it seems to be a valid id
with a slash at the end. In the latter case this also provides a link to
the potentially correct topic for convenience.

\Flow\Exception\InvalidTopicUuidException::__construct is only used in
that one single place, so changing the constructor signature should be
no problem.

The current implementation places the "Did you mean ...?"  message last
on the page, but that seems the easiest thing to do and good enough.
Changing the order might require more effort than we want to invest into
Flow.

Bug: T358616
Change-Id: Ia85e97fbed40c3fd843170c7f7eb4c32e5f4c77b