Remove broken/unused "stub post" system
We had something here that tried to display placeholders when post
data could not be loaded. Unfortunately it does not work: we just had
an outage because we broke the post loading code, and all this system
did was to generate more obscure exceptions than the ones we would
have seen otherwise (T364499).

Except for this outage, there are no entries in WMF production from
the two wfDebugLog() calls, so it seems it is also unused.

Remove it all in favor of throwing an exception.

Change-Id: I084274c228c8ad0f598064cff12558f02450faa4