Make RevisionFormatterTest run standalone
RevisionFormatterTest runs fine as part of the Flow test suite,
but fails when run as a standalone test class with PHPUnit. This
appears to be connected with lifecycle management for the
services container.

Turn the TalkpageManager into a service in both the Flow
container and the MediaWikiServices container so that the
lifecycle is properly managed.

Bug: T363442
Change-Id: I523ee219332562c33f7ccb35cc35a00a8a9aee66