Reset Hooks in testcase setup
The `Hooks` class of the `Flow` extension takes static references
to `$abuseFilter` and `$occupationController`. If these are not
reset between testcases, a test class may retain references to a
service container that has already been reset. This can result in
`Container is disabled!` error messages in PHPUnit tests.

Reset the Hooks at the start of every `PostRevisionTestCase`.

Bug: T368506
Bug: T368390
Change-Id: Ic09947a95fad377614f5ad2be6d4a4c4e27cc185