Drop extra reference to OccupationController service
This should not make any difference. Code that calls this method will
initialize the service. This was happening before and still happens.
But why do we need a local reference? Calling the upstream getService()
already takes care of that and will return the previously initialized
instance, if there is already one.

The problem with the extra local reference is that it might be
outdated, especially in test scenarios.

Bug: T368390
Change-Id: I6d2b728d8ada34224d877bffaa6c2ec37f5e2b85
(cherry picked from commit 20cbe0cd9df55f5cb0402f85b9085bd0ade5964a)