Use IReadableDatabase in more places, where possible
Direct follow-up for what we started in I1b23eae. In one case the
documentation explains that the database connection should be a
primary one. I'm not touching this documentation. Still all the code
actually does is reading. It's obviously not the responsibility of
this class to make sure that the same database connection is used
for reading and writing. It's technically not even possible for
this class to guarantee this. That's the responsibility of the
caller.

As well as:
* Make all @covers tags absolute.
* Delete a test that's just scaffolding but doesn't do anything.

Change-Id: I252854ccebf54e4338881edf00f920d9973fa5b0