Prepare Less styles for math=parens-division
Follows-up Ia14c0d8f82 (T116573, 317756cc61), and
Ie701a8fc3 (4f1e9d0081).

Note that border-radius actually supports a native `/` syntax, which
is part of why we are moving to math=parens-division soon, so that
this CSS feature can actually be used in Less.

However, if you open https://www.mediawiki.org/wiki/Manual_talk:FAQ/Flow?useskin=monobook
and inspect the input element, and set `border-radius: 0.25em/1.5`,
the browser will mark it as invalid and ignore it, because both
values must have a unit in order to use this feature. A border radius
of "1.5" has no defined meaning in CSS.

Currently this code is compiled by Less to a value of `0.16666667em`,
which is what shoudl stay here even after the meaning of slash is
changed from math=always to math=parens-division, as per Less.js v4
and Less.php v5.

Bug: T368921
Change-Id: Ice03d1b8bd8dde22b350672f9babc5d21e672957