build: Disable MediaWiki.WhiteSpace.SpaceBeforeSingleLineComment sniff
There are 49 files with 108 findings.

This extension is unmaintenanced and has many comments after code lines.
Sniffs listed in <exclude> get autofixed by libup and indicate that the
sniff should be fixed in the long term.
Disable the sniff with serverity = 0 to indicate that
this sniff does not need fixing in this extension.

Change-Id: I3993f35fa653fd31536b0b68aa175eed0894678f