api: die with error on unknown topic, not unhandled exception
This should take care of handling logstash logspam from people trying to
do something with a flow-topic that does not exist.

The error handling is not particularly nice, but this extension is
about to be removed everywhere in production.

The main cause of occurence for this error seems to be from people
clicking the example link on api-help pages like
https://www.mediawiki.org/w/api.php?action=help&modules=flow&recursivesubmodules=1
which directly triggers this exception.

Bug: T366027
Change-Id: If44c20cd745992a5acba3062a7f16f76cda6efd8