Add test to ensure that UUID JSON serialization is backwards-compatible
This is to guarantee that UUID JSON serialization stored in a database
can be deserialized by future versions of the application, and not
merely that the current version can round-trip the (de)serialization,
like in the existing test.

This is a "light" version of the procedure described at
https://www.mediawiki.org/wiki/Manual:Parser_cache/Serialization_compatibility
(which we use in MW core for serialization of more important things).

Bug: T391948
Change-Id: Id24cdf966fb2743ae6afe2c38677558b16caa92c