Do not try to import pages that already redirect to Flow
We've been making a plan to clean up ptwikibooks by first manually
redirecting LQT pages to the Flow topics they were imported to and then
running the standard import code. That code doesn't really support this
workflow, assuming its sourcestore logic is sufficient, so make it
support it.

This code is kind of in the wrong place - it's adding high-level
concepts to what is otherwise a low-level API, but it should work as a
quick fix for a codebase that's on it's way out anyway

Bug: T382435
Bug: T388099
Change-Id: I6ae8741e260128c085121ceac0eac2f0a5ccf300