LQT->Flow converter: Skip pages which have no threads
Now that we are considering Flow as an intermediate step
to converting these pages to plain talk pages, converting
an LQT board with no threads is at best pointless, and at
worst a regression (if a wikitext talk page is mistakenly
being treated as an LQT page).

Allow empty pages to still be converted by passing an
extra flag.

Change-Id: Idf5810f8ad4290d0fd3798aa5e2151d76fed421c