Prefer Title::newFromPageIdentity where possible
A page identity means we know the page id and can use it without
querying it again. A link target means we know only the page title
and need to query the page id again.

That's why this is more efficient in places where the page id is
already known.

Change-Id: I258b2b0c7e6d375c43693f788acddf6b019d75bd