Move some Echo hook handler to new class to make Echo optional
The migration of hook handlers for core hooks in 54926d9fd0
missed that the BeforeDisplayOrangeAlertHook is not in core,
but in Echo extension. This requires that Echo is always installed with
Flow as the interface is loaded and must exists.
Move this hook handler and two others to a new class to remove the
dependency.

Bug: T338450
Bug: T374325
Bug: T389998
Follow-Up: I5f809fd344d3e74846d9f140f5ab56685033ece1
Change-Id: I1604914fef6b86c82a35af22aa0a73e91eba7d16