Hackaround 2015 broken convert on ptwikibooks
Flow finds the title of the Flow board to add messages to by looking at
the move log by page ID (only considering Flow talk page manager edits).
For the specific page on ptwikibooks
https://pt.wikibooks.org/wiki/Wikilivros:Plant%C3%A3o_de_d%C3%BAvidas/Arquivo_LQT_1
it was accidentally double-converted to a "/Arquivo_LQT_1/Arquivo_LQT_1"
title (T117514), and the log entry from the double-convert is the one selected. Ignore entries where the "from" title is the same as the current title; it never makes sense for a LQT page to have been moved from itself.

Bug: T402549
Change-Id: I7c684ba4282a23b8cef840b9b70bd187c35f18ad